create user_prompt table row when inference call begins
